backend/requirements.txt
argcomplete
asgiref
Babel
bcrypt
blinker
certifi
chardet
click
distlib
distro
django
django-cors-headers
django-environ
django-imagekit
django-mathfilters
django-phonenumber-field
django-rest-framework
djangorestframework-simplejwt
djoser
fasteners
filelock
future
gunicorn
httplib2
idna
importlib-metadata
jeepney
Jinja2
keyring
launchpadlib
lazr.restfulclient
lazr.uri
livereload
lockfile
macaroonbakery
Mako
Markdown
MarkupSafe
mkdocs
monotonic
more-itertools
netifaces
oauthlib
olefile
packaging
paramiko
pexpect
phonenumbers
Pillow
pipx
platformdirs
protobuf
psutil
psycopg2
psycopg2-binary
ptyprocess
pydub
Pygments
pyinotify
PyJWT
pymacaroons
PyNaCl
pyparsing
pyRFC3339
pytz
pyxdg
reportlab
requests
SecretStorage
six
sqlparse
ssh-import-id
social-auth-app-django
stripe
style
tornado
update
urllib3
userpath
virtualenv
wadllib
wheel
xdg
zipp
